fix pca datatype

+++ fgsim/loaders/calochallange/pca.py
@@ -43,5 +43,5 @@ def fpc_from_batch(batch: Batch) -> dict[str, torch.Tensor]:
     # ].reshape(1, 3)
     # assert ((untrfs-fct).std(0)<(untrfs).std(0)).all()
     return dict(zip(["x", "y", "z"], first_pc.T.float().abs())) | {
-        "eval": e_vals[:, -1]
+        "eval": e_vals[:, -1].float()
     }
